requirements.txt
Eel==0.14.0
geopandas==0.10.2
matplotlib==3.3.1
mplleaflet==0.0.5
numpy~=1.21.4
Pillow==8.4.0
rdp==0.8
scikit-image==0.18.1
Fiona==1.8.20
Shapely==1.8.0
pytorch-lightning==1.5.0
pytorch-ranger==0.1.1
pytorch-toolbelt==0.4.4
torch==1.10.0
torch-optimizer==0.0.1a16
torchmetrics==0.6.0
torchsummary==1.5.1
torchvision==0.11.1
rtree
